+++ src/core/client/startup.ts
@@ -36,6 +36,7 @@ import './systems/toolbar';
 import './systems/vehicle';
 import './systems/world';
 // Utility
+import './utility/getters';
 import './utility/notification';
 import './utility/reload';
 import './utility/screenshot';